+++ src/pages/NwbPage/plugins/registry.ts
@@ -20,7 +20,11 @@ import { intervalSeriesPlugin } from "./IntervalSeries";
 import { eventsPlugin } from "./Events";
 import { imageSeriesMp4Plugin } from "./ImageSeriesMp4";
 import { NwbFileSpecifications } from "../SpecificationsView/SetupNwbFileSpecificationsProvider";
-import figpackRasterPlotPlugin, { figpackVideoPreviewPlugin } from "./Figpack";
+import {
+  FigpackPoseEstimationPlugin,
+  figpackVideoPreviewPlugin,
+  figpackRasterPlotPlugin,
+} from "./Figpack";
 
 // List of plugins in order they will appear in the UI when a single object is being viewed
 export const nwbObjectViewPlugins: NwbObjectViewPlugin[] = [
@@ -46,6 +50,7 @@ export const nwbObjectViewPlugins: NwbObjectViewPlugin[] = [
 
   figpackRasterPlotPlugin,
   figpackVideoPreviewPlugin,
+  FigpackPoseEstimationPlugin,
 
   defaultPlugin,
   pythonScriptPlugin,